Pendulum Class
For context: the main class is not working so I'm importing everyone's classes to this branch for it to work together.

This pendulum simulation is a rework of Anish's old pendulum model that was done as a joint collaboration between Me (Marco) and Anish. The contribution is about 50% for each (rough estimate).